# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CostConfig": {
        "properties": {
          "commission_bps": {
            "title": "Commission Bps",
            "type": "number"
          },
          "exchange_fee_bps": {
            "title": "Exchange Fee Bps",
            "type": "number"
          },
          "fixed_fee_per_fill": {
            "title": "Fixed Fee Per Fill",
            "type": "number"
          }
        },
        "title": "CostConfig",
        "type": "object"
      },
...
      "ExperimentDetailOut": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "costs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CostBreakdownOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "dataset_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Checksum"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when this response reuses an earlier experiment with an identical configuration and dataset checksum, instead of re-running the engine (Step 6 duplicate-submission policy). Refers to the experiment id whose result is being reused.",
            "title": "Duplicate Of"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpactEstimateOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "symbol",
          "dataset",
          "side",
          "quantity",
          "strategy",
          "status",
          "config"
        ],
        "title": "ExperimentDetailOut",
        "type": "object"
      },
      "ExperimentRequest": {
        "description": "POST /experiments body.\n\nstrategy-specific fields are only required for their own strategy;\nvalidated in `check_strategy_fields` below rather than left to fail\ndeep inside the engine with a confusing error.",
        "properties": {
          "arrival_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Price"
          },
          "costs": {
            "$ref": "#/components/schemas/CostConfig"
          },
          "dataset": {
            "description": "Path to a CSV dataset under datasets/, e.g. datasets/sample_synthetic.csv",
            "title": "Dataset",
            "type": "string"
          },
          "impact": {
            "$ref": "#/components/schemas/ImpactConfig"
          },
          "latency_ms": {
            "default": 0,
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "max_order_qty": {
            "default": 0,
            "minimum": 0.0,
            "title": "Max Order Qty",
            "type": "integer"
          },
          "min_order_qty": {
            "default": 1,
            "minimum": 1.0,
            "title": "Min Order Qty",
            "type": "integer"
          },
          "participation_rate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participation Rate"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "side": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Side",
            "type": "string"
          },
          "slices": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slices"
          },
          "strategy": {
            "enum": [
              "TWAP",
              "VWAP",
              "POV"
            ],
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "volume_profile": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Profile"
          }
        },
        "required": [
          "symbol",
          "dataset",
          "side",
          "quantity",
          "strategy"
        ],
        "title": "ExperimentRequest",
        "type": "object"
      },
      "ExperimentSummaryOut": {
        "properties": {
          "average_execution_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Execution Price"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "dataset_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Checksum"
          },
          "fill_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Rate"
          },
          "filled_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Quantity"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "requested_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Quantity"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "slippage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "total_cost_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost Bps"
          }
        },
        "required": [
          "id",
          "created_at",
          "symbol",
          "dataset",
          "side",
          "quantity",
          "strategy",
          "status"
        ],
        "title": "ExperimentSummaryOut",
        "type": "object"
      },
      "FillOut": {
        "properties": {
          "ask": {
            "title": "Ask",
            "type": "number"
          },
          "bid": {
            "title": "Bid",
            "type": "number"
          },
          "market_volume": {
            "title": "Market Volume",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "timestamp_ms": {
            "title": "Timestamp Ms",
            "type": "integer"
          }
        },
        "required": [
          "seq",
          "timestamp_ms",
          "price",
          "qty",
          "bid",
          "ask",
          "market_volume"
        ],
        "title": "FillOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImpactConfig": {
        "description": "Enable the simplified market-impact estimate (spec section 31). Off\nby default - it's an educational, non-fitted estimate, not a core\nexecution-quality metric, so it should be opt-in.",
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "eta": {
            "default": 0.1,
            "title": "Eta",
            "type": "number"
          }
        },
        "title": "ImpactConfig",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only + experiment-triggering API over the C++ execution engine.\n\n**Every result originates from an actual `ExecutionSession` run** (see `backend/experiment_service.py`) - there are no mocked responses (spec \u00a732/\u00a740). This is a research/backtesting system, **not** a live trading platform or broker (spec \u00a748).\n\nInteractive docs: **`/docs`** (Swagger UI) and **`/redoc`** (ReDoc). The raw schema is at **`/openapi.json`**; a committed copy lives at `docs/openapi.json` (regenerate with `scripts/export_openapi.py`).",
    "license": {
      "name": "See repository LICENSE"
    },
    "summary": "Historical market-execution simulator: run TWAP/VWAP/POV parent orders against replayed liquidity and read the resulting execution-quality analytics.",
    "title": "QuantExec Execution Engine API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/experiments": {
      "get": {
        "operationId": "list_experiments_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExperimentSummaryOut"
                  },
                  "title": "Response List Experiments Experiments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List past experiments (most recent first)",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Run one experiment through the real engine.\n\nStep 6 duplicate-submission policy: if an experiment with an\nidentical request body AND an identical dataset checksum already\ncompleted successfully, that result is returned as-is (with\n`duplicate_of` set) instead of re-running the engine - the same\ninputs are guaranteed to produce the same outputs (see BASELINE.md\n\u00a73.4), so re-running would only burn time for an identical answer.\nPass `?allow_duplicate=true` to force a fresh run anyway (e.g. to\nverify determinism, or after fixing something the config_hash can't see\nlike the engine binary itself).",
        "operationId": "create_experiment_experiments_post",
        "parameters": [
          {
            "in": "query",
            "name": "allow_duplicate",
            "required": false,
            "schema": {
              "default": false,
              "title": "Allow Duplicate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run an experiment through the engine",
        "tags": [
          "experiments"
        ]
      }
    },
    "/experiments/{experiment_id}": {
      "delete": {
        "description": "Step 6 'cancelled' state.\n\nHonest limitation: experiment execution in this backend is synchronous\n(the engine runs to completion inside the POST /experiments request\nhandler before it returns), so there is no in-flight run for this\nendpoint to interrupt under normal operation - by the time a client\ncould call DELETE, POST has already returned 'completed' or 'failed'.\nThis endpoint exists for the one case a status can legitimately still\nbe 'running': the server process crashed or was killed mid-request,\nleaving a row stuck in 'running' with no execution behind it. Marking\nthat row 'cancelled' is what unblocks a duplicate-submission retry of\nthe same config (see the config_hash lookup above, which only matches\n'completed' rows). A future async/queued execution mode is what would\nmake this endpoint able to interrupt a genuinely in-flight run.",
        "operationId": "cancel_experiment_experiments__experiment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a stranded 'running' experiment",
        "tags": [
          "experiments"
        ]
      },
      "get": {
        "operationId": "get_experiment_experiments__experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get one experiment's full detail",
        "tags": [
          "experiments"
        ]
      }
    },
    "/experiments/{experiment_id}/fills": {
      "get": {
        "operationId": "get_experiment_fills_experiments__experiment_id__fills_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FillOut"
                  },
                  "title": "Response Get Experiment Fills Experiments  Experiment Id  Fills Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get an experiment's individual fills",
        "tags": [
          "experiments"
        ]
      }
    },
...
  },
...
}